+++ IBAnimatable/AnimatorFactory.swift
@@ -36,6 +36,10 @@ struct AnimatorFactory {
       return SystemFlipAnimator(fromDirection: .Top, transitionDuration: transitionDuration)
     case .SystemFlipFromBottom:
       return SystemFlipAnimator(fromDirection: .Bottom, transitionDuration: transitionDuration)
+    case .SystemPageCurlFromTop:
+      return SystemPageCurlAnimator(fromDirection: .Top, transitionDuration: transitionDuration)
+    case .SystemPageCurlFromBottom:
+      return SystemPageCurlAnimator(fromDirection: .Bottom, transitionDuration: transitionDuration)
     }
   }
 }