+++ CHANGELOG.md
@@ -13,6 +13,7 @@ TBD
 
 - Support Xcode 7.3 and Swift 2.2
 - Configurable mask polygon (sides)
+- Add `SystemPageCurlAnimator` to support `SystemPageCurlFromTop` and `SystemPageCurlFromBottom` transition animations
 
 #### Bugfixes